Markup percentage and fees are working correctly now

+++ src/fr/ecp/IS1220/group5/project/Financial.java
@@ -1,6 +1,7 @@
 package fr.ecp.IS1220.group5.project;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 /**
  * <b>The class that stores the financial variables</b>
@@ -9,8 +10,8 @@ public class Financial implements Serializable {
 
 	private static final long serialVersionUID = 3789197273932947254L;
 
-	public double service_fee;
-	public double markup_percentage;
-	public double delivery_cost;
+	public BigDecimal service_fee;
+	public BigDecimal markup_percentage;
+	public BigDecimal delivery_cost;
 	
 }
